update Run() and RunGetTags() methods
these methods need to be updated because the interface slightly changed.
the biggest change is that filters now need to be prepared before they
can be applied, this can be done in parallel with the getting of the
initial ids. once the filters are prepared then we can start applying
them to the initial ids.